// frontend/src/app/core/api/models/thesisDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ThesisDto { 
    id?: string;
    studentId?: string;
    fileName?: string;
    status?: ThesisDto.StatusEnum;
    rejectionMessage?: string;
    uploadedAt?: string;
    checkedAt?: string;
}
export namespace ThesisDto {
    export type StatusEnum = 'UNCHECKED' | 'CHECKED' | 'REJECTED';
    export const StatusEnum = {
        Unchecked: 'UNCHECKED' as StatusEnum,
        Checked: 'CHECKED' as StatusEnum,
        Rejected: 'REJECTED' as StatusEnum
    };
}